+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Remove package dependency on libseccomp.
 * Added detection of libnvdxgdmal.so.1 on WSL2.
 * Fix bug in determining default nvidia-container-runtime.user config value on SUSE-based systems.
+* Add `crun` to the list of configured low-level runtimes.
 
 * [toolkit-container] Bump CUDA base image version to 12.3.1.