+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Add `crun` to the list of configured low-level runtimes.
 
 * [toolkit-container] Bump CUDA base image version to 12.3.1.
+* [libnvidia-container] Added detection of libnvdxgdmal.so.1 on WSL2.
 
 ## v1.14.3
 * [toolkit-container] Bump CUDA base image version to 12.2.2.